fix scroll to anchor
